fix: validations, console errors and some styles
