+++ src/app/(pages)/drugs-wizard/context.jsx
@@ -1,0 +1,3 @@
+import { createContext } from 'react';
+
+export const WizardContext = createContext("");
