+++ src/app/(pages)/drugs-wizard/step1/page.jsx
@@ -1,52 +1,61 @@
 'use client'
-
 import { useForm, FormProvider } from "react-hook-form"
-import { useContext } from "react";
-import { WizardContext } from "../wizard-context";
-import MainForm from "@/app/components/main-form";
+import { useContext, useEffect } from "react";
+import MainForm from "../../../components/main-form";
 import { Box, Button } from "@mui/material";
-import asas from "../../../constants/asas.json";
 import { useRouter } from "next/navigation";
-import NavigateNextIcon from '@mui/icons-material/NavigateNext';
+import { NavigateNext } from '@mui/icons-material';
+import { WizardContext } from "../context";
+import asas from "../../../constants/asas.json";
+import Footer from "../../../components/footer";
 
 export default function DrugWizardStep1() {
-    const mainForm = useForm({ mode: "onBlur" });
-    const { watch, handleSubmit } = mainForm;
+  const mainForm = useForm({ mode: "onBlur" });
+  const { setValue } = mainForm;
+  const { watch, handleSubmit } = mainForm;
 
-    const { asaObj, weightObj } = useContext(WizardContext);
-    const [, setWeightState] = weightObj;
-    const [selectedAsa, setSelectedAsa] = asaObj;
+  const { asaObj, weightObj } = useContext(WizardContext);
+  const [weightValue, setWeightState] = weightObj;
+  const [selectedAsa, setSelectedAsa] = asaObj;
 
-    const router = useRouter()
+  const router = useRouter()
+  useEffect(() => {
+    setValue('weight', weightValue);
+  }, []);
 
-    const { weight } = watch();
+  const { weight } = watch();
+  setWeightState(weight);
+      
+  const { asa } = watch();
+  const filteredSelectedAsa = asas.find(a => a.id === asa);
+  setSelectedAsa(filteredSelectedAsa);
 
-    setWeightState(weight)
-        
-    const { asa } = watch();
-    const filteredSelectedAsa = asas.find(a => a.id === asa);
-    setSelectedAsa(filteredSelectedAsa);
+  const { formState: { isValid } } = mainForm;
 
-
-    return (
-        <>
-            <FormProvider {...mainForm}>
-                <MainForm asa={selectedAsa} disableWeight={false} />
-            </FormProvider>
-            <Box
-                sx={{
-                    marginTop: 3,
-                    display: 'flex',
-                    justifyContent: 'flex-end'
-
-                }}
-            >
-                <Button
-                    startIcon={<NavigateNextIcon />}
-                    variant="contained" onClick={handleSubmit(() => router.push('./step2'))} >
-                    Siguiente
-                </Button>
-            </Box>
-        </>
-    )
+  return (
+    <>
+      <FormProvider {...mainForm}>
+        <MainForm asa={selectedAsa} disableWeight={false} />
+      </FormProvider>
+      <Box
+        sx={{
+          marginTop: 3,
+          display: 'flex',
+          justifyContent: 'flex-end'
+        }}
+      >
+        <Footer>
+          <Button
+            endIcon={<NavigateNext />}
+            variant="contained"
+            onClick={handleSubmit(() => router.push('./step2'))}
+            disabled={!isValid}
+            fullWidth={true}
+          >
+            Siguiente
+          </Button>
+        </Footer>
+      </Box>
+    </>
+  )
 }
