+++ src/app/(pages)/drugs-wizard/step4/page.jsx
@@ -1,53 +1,54 @@
 
 'use client'
-import { WizardContext } from "../wizard-context";
+import { WizardContext } from "../context";
 import { useContext } from "react";
-import { Box, Button } from "@mui/material";
-import Budget from "@/app/components/budget";
-import CleaningServicesIcon from '@mui/icons-material/CleaningServices';
+import { Box, Button, ButtonGroup } from "@mui/material";
+import Budget from "../../../components/budget";
 import { useRouter } from "next/navigation";
+import Footer from "../../../components/footer";
+import { NavigateBefore } from "@mui/icons-material";
 
 export default function DrugWizardStep4() {
-    const { asaObj, drugObj, weightObj } = useContext(WizardContext);
-    const [weightState, setWeightState] = weightObj;
-    const [selectedAsa, setSelectedAsa] = asaObj;
-    const [drugs, setDrugs] = drugObj;
-    const router = useRouter();
-    
-    const limpiar = () => {
-        setSelectedAsa();
-        setDrugs([]);
-        setWeightState();
-        router.push('./step1')
-    }
+  const { asaObj, drugObj, weightObj } = useContext(WizardContext);
+  const [weightState, setWeightState] = weightObj;
+  const [selectedAsa, setSelectedAsa] = asaObj;
+  const [drugs, setDrugs] = drugObj;
+  const router = useRouter();
+  
+  const goToStep1 = () => {
+    setSelectedAsa();
+    setDrugs([]);
+    setWeightState();
+    router.push('./step1')
+  }
 
-    return (
-        <Box>
-            <Budget
-                asa={selectedAsa}
-                weight={weightState}
-                drugs={drugs}
-                disabled={!drugs.length}
-            />
-            <Box
-                sx={{
-                    marginTop: 3,
-                    display: 'flex',
-                    justifyContent: 'flex-end'
-
-                }}
-            >
-                <Button
-                    startIcon={<CleaningServicesIcon />}
-                    sx={{ marginX: 1 }}
-                    color="error"
-                    variant="contained" onClick={() => limpiar()} >
-                    Limpiar
-                </Button>
-                <Button sx={{ marginX: 1 }} variant="contained" onClick={() => router.push('./step3')} >
-                    Volver
-                </Button>
-            </Box>
-        </Box>
-    )
+  return (
+    <Box>
+      <Budget
+        asa={selectedAsa}
+        weight={weightState}
+        drugs={drugs}
+        disabled={!drugs.length}
+      />
+      <Footer>
+        <ButtonGroup fullWidth={true}>
+          <Button
+            sx={{ marginX: 1 }}
+            startIcon={<NavigateBefore />}
+            variant="outlined"
+            onClick={() => router.push('./step3')}
+          >
+            Volver
+          </Button>
+          <Button
+            sx={{ marginX: 1 }}
+            variant="contained"
+            onClick={() => goToStep1()}
+          >
+            Nuevo
+          </Button>
+        </ButtonGroup>
+      </Footer>
+    </Box>
+  )
 }
