--- src/app/(pages)/drugs-wizard/wizard-context.js
@@ -1,3 +1,0 @@
-import { createContext } from 'react';
-
-export const WizardContext = createContext("");
