+++ src/app/components/drug-list/index.jsx
@@ -1,8 +1,10 @@
 import List from '@mui/material/List';
 import DrugItem from './drug-item';
+import { Divider } from '@mui/material';
 
 const DrugList = ({ list = [], handleOnDelete }) => (
   <List dense={true}>
+    <Divider />
     {list.map((row) => <DrugItem key={row.id} drug={row} handleOnDelete={handleOnDelete} />)}
   </List>
 );
