+++ src/app/components/footer/style.module.css
@@ -1,0 +1,8 @@
+.footer {
+  position: fixed;
+  left: 0;
+  bottom: 0;
+  width: 100%;
+  color: white;
+  text-align: center;
+}
