+++ src/app/components/nav-bar/links.jsx
@@ -1,29 +1,29 @@
-import HomeIcon from '@mui/icons-material/Home';
+// import HomeIcon from '@mui/icons-material/Home';
 import VaccinesIcon from '@mui/icons-material/Vaccines';
-import SettingsIcon from '@mui/icons-material/Settings';
-import ContentCopyIcon from '@mui/icons-material/ContentCopy';
-export const Links =  [
-{
+// import SettingsIcon from '@mui/icons-material/Settings';
+// import ContentCopyIcon from '@mui/icons-material/ContentCopy';
+
+export const Links = [
+  /* {
     text: 'Home',
     path: '/',
     icon: <HomeIcon />
-},
-{
+  }, */
+  {
     text: 'Wizard',
     path: '/drugs-wizard/step1',
     icon: <VaccinesIcon />
-},
-{
-
+  },
+  /* {
     text: 'Configuraciones',
     path: '/config',
     icon: <SettingsIcon />
-},
-{
+  },
+  {
     text: 'Plantillas',
     path: '/templates',
     icon: <ContentCopyIcon />
-},
+  }, */
 ]
